fix: twotone name issue for d.ts and .js in alpha.1
